[AND-279] Move Promotional FG image to promotional feature folder
