+++ app-games/src/main/java/com/aptoide/android/aptoidegames/feature_promotional/EventPromotionalView.kt
@@ -23,7 +23,6 @@ import cm.aptoide.pt.feature_home.domain.Bundle
 import cm.aptoide.pt.feature_home.domain.randomBundle
 import com.aptoide.android.aptoidegames.R
 import com.aptoide.android.aptoidegames.editorial.buildEditorialRoute
-import com.aptoide.android.aptoidegames.home.AptoidePromotionalFeatureGraphicImage
 import com.aptoide.android.aptoidegames.home.LoadingBundleView
 import com.aptoide.android.aptoidegames.theme.AGTypography
 import com.aptoide.android.aptoidegames.theme.AptoideTheme
